+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.0] - 2021-09-24
 ### Added
+- Add the `isActivated` var for the `Campaign` entity.
+- Add the `getActiveCampaign()` method in the `CampaignsClient` class
 - Added tests for Campaigns entities and for some methods of the client
 
 ### Changed